backend/requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
mysql-connector-python
pyjwt
